+++ django_goodies/models/__init__.py
@@ -1,3 +1,4 @@
 from django_goodies.models.models import (  # NOQA
     ArchivableMixin, CommonInfoMixin, StaticAbstract, VersioningMixin
 )
+from django_goodies.models.fields import TimeZoneField  # NOQA